Remade db schema table
